requirements.txt
# Required libraries for DevMind
# Install with: pip install -r requirements.txt

anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
